pyproject.toml
[tool.poetry]
name = "shotstriper"
version = "0.1.0"
description = "Add beautiful, striped backgrounds to your screenshots"
authors = ["Helmut Irle <me@helmut.dev>"]
readme = "README.md"

[tool.poetry.scripts]
shotstriper = "shotstriper.main:app"

[tool.poetry.dependencies]
python = "^3.10"
typer = {extras = ["all"], version = "^0.10.0"}
pillow = "^10.2.0"
aggdraw = "^1.3.18.post0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
